glt - Add new UCSB API method to UCSBCurriculumService
For a particular course, we need a way to find the primaries and
secondaries associated with that course.  The existing method
getSection accesses a UCSB API endpoint that only gives the
information about the section with the enroll code argument, and none
of the related sections.

Importantly, there is no way to find a primary from the secondary at
this endpoint.  Therefore, we have added a new method for another UCSB
API endpoint that provides this feature in order to efficiently get
the primaries and secondaries related to an enroll code.

It may have been possible to replace the old endpoint, but I don't
know if that would break any other code in the project.