pc - filter out weird subjectArea records
Sometimes we end up with weird records in the subjectArea table that have field values
that are identical to the field names.  We are not sure if this is a fault of the API,
or a consequence of some other bug.  It hasn't happened in a while, so it may have been
a transient bug in the UCSB Developer API, and this PR may no longer be neeeded.

But we'll leave this here as a draft PR for now in case it arises again.