pc - make default for exception be blank
The two swagger endpoints for CSV downloads have a parameter where you can pass in
the name of an exception that you want to be thrown so that you can test the
exception handling code.

This change makes those blank so that folks using the swagger endpoint
don't have to clear out that field when they are testing the normal case.