pc - modify Dockerfile so that the build output shows errors more clearly
In the Dockerfile output for the mvn package step, there are
frequently so many messages for 'Downloading such and such' that
clutter up the output, and make it hard to find the *real*
significant errors.  This commit adds a couple of extra flags
that supress output that is not really that useful, so that
build errors are easier to find.