pc - add backend endpoint that returns a list of primaries
This is a first step towards simplifying the structure of the frontend code
for displaying courses.  The new TanStack react table has much nicer features
for condensing rows that can work with the JSON without having to do the
tricks we had to do before with flattening the sections (the "ConvertedSection" thing).

This should make things much easier; we can *almost* work with the JSON directly
in the form in which it comes from the UCSB API.  We only have to make one small tweak,
which is to separate out the individual primaries into separate objects.